优化completableFuture 测试案例

+++ demo/src/main/java/com/future/Quote.java
@@ -1,5 +1,7 @@
 package com.future;
 
+import java.math.BigDecimal;
+
 /**
  * @author louis
  * <p>
@@ -9,9 +11,9 @@ package com.future;
 public class Quote {
 
     private final String shopName;
-    private final double price;
+    private final BigDecimal price;
     private final Discount.Code discountCode;
-    public Quote(String shopName, double price, Discount.Code code) {
+    public Quote(String shopName, BigDecimal price, Discount.Code code) {
         this.shopName = shopName;
         this.price = price;
         this.discountCode = code;
@@ -19,14 +21,11 @@ public class Quote {
     public static Quote parse(String s) {
         String[] split = s.split(":");
         String shopName = split[0];
-        double price = Double.parseDouble(split[1]);
+        BigDecimal price = BigDecimal.valueOf(Double.parseDouble(split[1]));
         Discount.Code discountCode = Discount.Code.valueOf(split[2]);
         return new Quote(shopName, price, discountCode);
     }
     public String getShopName() { return shopName; }
-    public double getPrice() { return price; }
+    public BigDecimal getPrice() { return price; }
     public Discount.Code getDiscountCode() { return discountCode; }
-
-
-
 }
